pyproject.toml
[project]
name = "slackbotbuilder"
version = "1.0.0"
description = "AI Slack Bot Builder with FastAPI"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "PyGithub>=1.59.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "PyYAML>=6.0.1",
    "openai==1.97.0",
    "tiktoken>=0.7.0",
    "slack-sdk>=3.36.0",
    "markdown-to-mrkdwn>=0.2.0",
]
requires-python = ">=3.8.1"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0,<7.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = [
    "*.py",
    "*.yaml",
    "*.json",
    "scripts/*.py",
]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"] 
